Trigger PlaceDao to get the places for the given trip

+++ app/src/main/java/com/gowthamraj07/journeytracker/data/db/dao/PlaceDao.kt
@@ -8,5 +8,5 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface PlaceDao {
     @Query("SELECT * FROM places WHERE tripId = :tripId")
-    fun getPlacesByTrip(tripId: Long): Flow<List<PlaceEntity>>
+    fun getPlacesByTrip(tripId: Int): Flow<List<PlaceEntity>>
 }
